+++ masterbase/app.py
@@ -117,7 +117,8 @@ def late_bytes(request: Request, api_key: str, data: LateBytesBody) -> dict[str,
     engine = request.app.state.engine
     current_time = datetime.now().astimezone(timezone.utc)
     steam_id = steam_id_from_api_key(engine, api_key)
-    late_bytes_helper(engine, steam_id, data.converted_late_bytes, current_time)
+    converted_late_bytes = bytes.fromhex(data.late_bytes)
+    late_bytes_helper(engine, steam_id, converted_late_bytes, current_time)
 
     return {"late_bytes": True}